// frontend/src/app/core/api/generated/api/suppliers.service.ts
export interface FindAll2RequestParams {
    page?: number;
    size?: number;
    code?: string;
    name?: string;
    fiscalIdentifier?: string;
    active?: boolean;
}